requirements.txt
requests
beautifulsoup4
pymupdf
google-cloud-storage
python-dotenv
chromadb
google-genai
fastapi
uvicorn
